requirements.txt
easyocr==1.6.2
imutils==0.5.4
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.2.0
PyQt5==5.15.7
